requirements.txt
alembic==1.7.6
Babel==2.9.1
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dnspython==2.2.0
dominate==2.6.0
elastic-transport==8.1.0
elasticsearch==7.13.3
email-validator==1.1.3
Flask==2.0.3
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
langdetect==1.0.9
Mako==1.1.6
MarkupSafe==2.1.0
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.1
